Minor changes - naming, code-splitting ...
